pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "trdr"
version = "0.0.1"
authors = [
    { name = "developer_jeb", email = "developer_jeb@outlook.com" },
]
description = "Trading strategy parser and executor"
dependencies = [
    "yfinance==0.2.51",
    "aiohttp==3.11.11",
    "opentelemetry-api==1.29.0",
    "opentelemetry-sdk==1.29.0",
    "pydantic==2.10.6",
]

[project.optional-dependencies]
dev = [
    "pytest==8.3.4",
    "pytest-asyncio==0.25.2",
    
]

[tool.pytest.ini_options]
pythonpath = ["src"]
